Fix html syntax by moving edit inside item
