+++ client/extensions/woocommerce/app/settings/payments/payment-method-list.js
@@ -9,90 +9,45 @@ import { localize } from 'i18n-calypso';
 /**
  * Internal dependencies
  */
-import {
-	changePaymentMethodField,
-	closeEditingPaymentMethod,
-	openPaymentMethodForEdit,
-} from 'woocommerce/state/ui/payments/methods/actions';
-import { errorNotice, successNotice } from 'state/notices/actions';
-import { fetchPaymentMethods, savePaymentMethod } from 'woocommerce/state/sites/payment-methods/actions';
-import { getCurrentlyEditingPaymentMethod, getPaymentMethodsGroup } from 'woocommerce/state/ui/payments/methods/selectors';
+import { fetchPaymentMethods } from 'woocommerce/state/sites/payment-methods/actions';
+import { getPaymentMethodsGroup } from 'woocommerce/state/ui/payments/methods/selectors';
 import { getSelectedSiteWithFallback } from 'woocommerce/state/sites/selectors';
 import List from 'woocommerce/components/list/list';
 import ListHeader from 'woocommerce/components/list/list-header';
+import ListItem from 'woocommerce/components/list/list-item';
 import ListItemField from 'woocommerce/components/list/list-item-field';
-import PaymentMethodEdit from './payment-method-edit';
 import PaymentMethodItem from './payment-method-item';
 
 class SettingsPaymentsMethodList extends Component {
 	static propTypes = {
-		closeEditingPaymentMethod: PropTypes.func.isRequired,
-		currentlyEditingMethod: PropTypes.shape( {
-			id: PropTypes.string,
-		} ),
 		fetchPaymentMethods: PropTypes.func.isRequired,
 		methodType: PropTypes.string.isRequired,
-		openPaymentMethodForEdit: PropTypes.func.isRequired,
 		paymentMethods: PropTypes.array.isRequired,
-		savePaymentMethod: PropTypes.func.isRequired,
 		site: PropTypes.object,
 	};
 
-	componentDidMount() {
-		this.props.fetchPaymentMethods();
-	}
-
-	onCancel = ( method ) => {
-		this.props.closeEditingPaymentMethod( this.props.site.ID, method.id );
-	}
-
-	onEdit = ( method ) => {
+	componentDidMount = () => {
 		const { site } = this.props;
-		this.props.openPaymentMethodForEdit( site.ID, method.id );
+
+		if ( site && site.ID ) {
+			this.props.fetchPaymentMethods( site.ID );
+		}
 	}
 
-	onEditField = ( field, value ) => {
-		this.props.changePaymentMethodField( this.props.site.ID, field, value );
-	}
+	componentWillReceiveProps = ( newProps ) => {
+		const { site } = this.props;
 
-	onSave = ( method ) => {
-		const { site, translate } = this.props;
+		const newSiteId = newProps.site && newProps.site.ID || null;
+		const oldSiteId = site && site.ID || null;
 
-		const successAction = () => {
-			this.props.closeEditingPaymentMethod( site.ID, method.id );
-			return successNotice(
-				translate( 'Payment method successfully saved.' ),
-				{ duration: 4000 }
-			);
-		};
-
-		const errorAction = () => {
-			return errorNotice(
-				translate( 'There was a problem saving the payment method. Please try again.' )
-			);
-		};
-
-		this.props.savePaymentMethod( site.ID, method, successAction, errorAction );
+		if ( oldSiteId !== newSiteId ) {
+			this.props.fetchPaymentMethods( newSiteId );
+		}
 	}
 
 	renderMethodItem = ( method ) => {
-		const currentlyEditingId = this.props.currentlyEditingMethod &&
-			this.props.currentlyEditingMethod.id;
-		const methodTypeClass = `payments__methods-${ this.props.methodType }`;
 		return (
-			<div className={ methodTypeClass } key={ method.title }>
-				<PaymentMethodItem
-					currentlyEditingId={ currentlyEditingId }
-					method={ method }
-					onCancel={ this.onCancel }
-					onEdit={ this.onEdit } />
-				{ currentlyEditingId === method.id && (
-					<PaymentMethodEdit
-						method={ this.props.currentlyEditingMethod }
-						onEditField={ this.onEditField }
-						onSave={ this.onSave } />
-				) }
-			</div>
+			<PaymentMethodItem method={ method } key={ method.title } />
 		);
 	}
 
@@ -120,11 +75,9 @@ class SettingsPaymentsMethodList extends Component {
 }
 
 function mapStateToProps( state, ownProps ) {
-	const currentlyEditingMethod = getCurrentlyEditingPaymentMethod( state );
 	const paymentMethods = getPaymentMethodsGroup( state, ownProps.methodType );
 	const site = getSelectedSiteWithFallback( state );
 	return {
-		currentlyEditingMethod,
 		paymentMethods,
 		site,
 	};
@@ -133,11 +86,7 @@ function mapStateToProps( state, ownProps ) {
 function mapDispatchToProps( dispatch ) {
 	return bindActionCreators(
 		{
-			changePaymentMethodField,
-			closeEditingPaymentMethod,
 			fetchPaymentMethods,
-			openPaymentMethodForEdit,
-			savePaymentMethod,
 		},
 		dispatch
 	);
